fix(coupling): address PR review for ContinuousAffineCoupling
- Make `time_net` configurable via constructor arg (Callable[[ArrayLike], Array]);
  default builds TimeFourier as before so `TimeTanh` / `TimeIdentity` are now
  actually usable via this coupling layer rather than being export-only.
- Document in the docstring that a single layer leaves the passive half
  untouched and that callers must stack with alternating masks for full
  coverage.
- Drop `dtype=float` coercion in `pack_time_control` / `unpack_time_control` /
  `_scalar_time` so public helpers respect JAX's default dtype promotion.
- Simplify the final-layer weight/bias scaling: drop the dead `None`-bias
  branch since `eqx.nn.MLP` always materializes a bias.
- Remove redundant "shape arrow" comments; the named variables already
  communicate the contract.
- Tighten the training-sanity assertion with a measurable margin
  (`final_loss < initial_loss - 0.05`) so numerics drift can't silently flake
  the test.
- Add `test_continuous_affine_coupling_accepts_custom_time_net` covering the
  three shipped time gates as drop-in replacements.

+++ src/gauss_flows/_src/conditioning.py
@@ -13,7 +13,7 @@ def _validate_control_dim(control_dim: int) -> None:
 
 
 def _as_time_column(t: ArrayLike) -> Array:
-    t_array = jnp.asarray(t, dtype=float)
+    t_array = jnp.asarray(t)
     if t_array.ndim == 0:
         return jnp.reshape(t_array, (1,))
     if t_array.shape[-1] == 1:
@@ -49,7 +49,7 @@ def pack_time_control(
     if control is None:
         return t_array
 
-    control_array = jnp.atleast_1d(jnp.asarray(control, dtype=float))
+    control_array = jnp.atleast_1d(jnp.asarray(control))
     if t_array.shape[:-1] != control_array.shape[:-1]:
         raise ValueError(
             "t and control must share the same leading batch shape; "
@@ -81,7 +81,7 @@ def unpack_time_control(
         ((1,), (2,))
     """
     _validate_control_dim(control_dim)
-    condition = jnp.asarray(condition, dtype=float)
+    condition = jnp.asarray(condition)
     if condition.ndim == 0:
         raise ValueError("condition must have at least one dimension.")
